+++ packages/pcd/ethereum-group-pcd/test/EthereumGroupPCD.spec.ts
@@ -131,6 +131,7 @@ async function happyPathEthGroupPCD(
 }
 
 describe("Ethereum Group PCD", function () {
+  // eslint-disable-next-line no-restricted-syntax
   this.timeout(60 * 1000 * 10);
   let ethGroupPCD: EthereumGroupPCD;