+++ packages/pcd/input-test-pcd/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pcd/input-test-pcd
 
+## 0.11.0
+
+### Minor Changes
+
+- Lightweight PCD package imports and new authentication libraries
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/pcd-types@0.11.0
+
 ## 0.10.0
 
 ### Minor Changes