+++ packages/tools/perftest/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pcd/perftest
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/pcd-types@0.11.0
+  - @pcd/pod@0.1.0
+  - @pcd/util@0.5.0
+  - @pcd/eddsa-ticket-pcd@0.6.0
+  - @pcd/pod-pcd@0.1.0
+  - @pcd/semaphore-identity-pcd@0.11.0
+  - @pcd/semaphore-signature-pcd@0.11.0
+  - @pcd/zk-eddsa-event-ticket-pcd@0.5.0
+
 ## 0.0.1
 
 ### Patch Changes