+++ packages/ui/message-pcd-ui/CHANGELOG.md
@@ -2,4 +2,17 @@
 
 ## 0.1.0
 
+### Minor Changes
+
+- Lightweight PCD package imports and new authentication libraries
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/passport-ui@0.11.0
+  - @pcd/pcd-types@0.11.0
+  - @pcd/message-pcd@0.1.0
+
+## 0.1.0
+
 - Initial commit