+++ packages/ui/semaphore-signature-pcd-ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pcd/semaphore-signature-pcd-ui
 
+## 0.2.0
+
+### Minor Changes
+
+- Lightweight PCD package imports and new authentication libraries
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/passport-ui@0.11.0
+  - @pcd/pcd-types@0.11.0
+  - @pcd/semaphore-signature-pcd@0.11.0
+
 ## 0.1.0
 
 ### Minor Changes